Move nodes to src
