+++ src/parser.rs
@@ -7,7 +7,7 @@ use std::path::{Path, PathBuf};
 
 use crate::codegen::{ERR_STR, NOTE_STR, WARN_STR};
 use crate::checker::Type;
-use crate::new::nodes::{self, Expression};
+use crate::nodes::{self, Expression};
 
 // We always store the N-1 next tokens for lookahead purposes, even if we only use 1 right now
 const LOOKAHEAD_LIMIT: usize = 3;
